Rollback for record admin decision
