+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/TseAdmReplyService.java
@@ -18,7 +18,7 @@ import uk.gov.hmcts.et.common.model.ccd.types.DocumentType;
 import uk.gov.hmcts.et.common.model.ccd.types.RespondentSumType;
 import uk.gov.hmcts.et.common.model.ccd.types.TseRespondType;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.NotificationHelper;
-import uk.gov.hmcts.ethos.replacement.docmosis.helpers.TseAdminHelper;
+import uk.gov.hmcts.ethos.replacement.docmosis.helpers.TseAdmReplyHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.TSEAdminEmailRecipientsData;
 
 import java.time.LocalDate;
@@ -291,7 +291,7 @@ public class TseAdmReplyService {
         if (isEmpty(caseData.getDocumentCollection())) {
             caseData.setDocumentCollection(new ArrayList<>());
         }
-        DocumentTypeItem docItem = TseAdminHelper.getDocumentTypeItem(documentManagementService, tornadoService,
+        DocumentTypeItem docItem = TseAdmReplyHelper.getDocumentTypeItem(documentManagementService, tornadoService,
                 caseDetails, userToken, TSE_ADMIN_REPLY, TSE_ADMIN_CORRESPONDENCE);
         caseData.getDocumentCollection().add(docItem);
     }
